requirements.txt
pyes==0.90.1
boto==2.19.0
PyYAML==3.10
pika==0.9.13